refactor: rename components for authentication and project management
- Introduced AuthIllustration, GitHubButton, and LoginForm components to enhance the authentication experience.
- Added PinnedProjects, ProfileHero, and various skeleton components for improved profile and project displays.
- Implemented error handling components for better user feedback during data loading failures.
- Created project-related components including ProjectGrid, ProjectHero, and ProjectEditForm to facilitate project management.
- Updated import paths across the application to reflect the new component structure, ensuring consistency and modularity.
