+++ apps/web/src/features/auth/components/auth-illustration.component.tsx
