+++ apps/web/src/features/profile/components/skeletons/skeleton-pinned-projects.component.tsx
