+++ apps/web/src/features/projects/components/error-ui/homepage-error.component.tsx
