+++ apps/web/src/features/projects/components/error-ui/project-detail-content-error.component.tsx
