+++ apps/web/src/features/projects/components/filter-search-bar.component.tsx
