+++ apps/web/src/features/projects/components/project-grid.component.tsx
