+++ apps/web/src/features/projects/components/project-hero.component.tsx
