+++ apps/web/src/features/projects/components/project-search-bar.component.tsx
