+++ apps/web/src/features/projects/components/project-side-bar.component.tsx
