+++ apps/web/src/features/projects/components/skeletons/skeleton-project-card.component.tsx
