+++ apps/web/src/features/projects/components/stepper/step-one.component.tsx
