+++ apps/web/src/features/projects/components/stepper/step-three.component.tsx
