+++ apps/web/src/features/projects/components/stepper/step-two.component.tsx
