+++ apps/web/src/features/projects/components/stepper/stepper.component.tsx
