+++ apps/web/src/features/projects/views/project-edit.view.tsx
@@ -2,10 +2,10 @@
 
 import Breadcrumb from "@/shared/components/shared/Breadcrumb";
 
-import ProjectEditFormError from "../components/error-ui/project-edit-form-error";
-import ProjectEditForm from "../components/project-edit-form";
-import ProjectEditSidebar from "../components/project-edit-sidebar";
-import SkeletonProjectEditForm from "../components/skeletons/skeleton-project-edit-form";
+import ProjectEditFormError from "../components/error-ui/project-edit-form-error.component";
+import ProjectEditForm from "../components/project-edit-form.component";
+import ProjectEditSidebar from "../components/project-edit-sidebar.component";
+import SkeletonProjectEditForm from "../components/skeletons/skeleton-project-edit-form.component";
 import { useProject } from "../hooks/use-projects.hook";
 
 export default function ProjectEditView({ projectId }: { projectId: string }) {
